refactor(upload): move upload orchestration into ingestion pipeline service
